cpm - squash mismatch warnings

+++ include/cpm.h
@@ -24,7 +24,7 @@ extern void *_CLIB_OPEN_MAX;
 
 
 
-#define MAXFILE CLIB_OPEN_MAX
+#define MAXFILE (int)CLIB_OPEN_MAX
 
 /* If you want fileio to support devices then link with -lcpmdevice.
  * This adds the following devices: CON: RDR: PUN: LST: without this
